test(spawn): split spawn_tests.rs into feature-scoped sub-modules
Closes #1089. Phase 1 follow-up to #1088.

`spawn_tests.rs` was extracted in #1088 at 1412 lines, over the U-16
800-line ceiling. Codex flagged it as P1. This split groups tests by
the feature they exercise and lifts shared mocks into a `helpers`
sibling module so each test file stays focused.

New layout under `task_runner/spawn_tests/`:
- `mod.rs` (189 lines) — workspace cleanup, register_pending_task,
  preregistered_metadata_refresh, transient_error_detection,
  local_waiting_counter, plus shared `tid` helper
- `helpers.rs` (198 lines) — CapturingAgent, BlockingInterceptor,
  PhaseCapturingAgent, wait_until, wait_for_captured_phases,
  wait_for_captured_prompts (used by skills + resume tests)
- `spawn_skills_tests.rs` (401 lines) — skills_are_injected,
  blocking_interceptor_fails, spawn_blocking_panic,
  execution_phase_is_set, validation_phase_is_set
- `spawn_resume_tests.rs` (212 lines) — resumed_pr_* (manual conflict,
  clean conflict gate, rebase push fresh-review prompt)
- `spawn_classify_tests.rs` (149 lines) — TaskKind classify cases,
  parse_pr_url_*, prompt_requires_plan_*, is_issue_pr_task
- `spawn_decomposition_tests.rs` (278 lines) — non_decomposable
  prompt source gating, spawn_awaiting_*, check_awaiting_*

`spawn.rs` updated from `#[path = "spawn_tests.rs"]` to
`#[path = "spawn_tests/mod.rs"]`. Submodules use `super::super::*` to
reach into spawn's namespace.

Zero logic changes — every assertion preserved verbatim.

Verification:
- cargo test -p harness-server --lib -- task_runner::spawn:
    39/39 passed (same set as #1088 — covers all spawn::tests::*)
- cargo fmt --all -- --check: clean
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean

All sub-modules are well under the U-16 800-line ceiling.
